requirements.txt
fastapi
sqlalchemy
uvicorn
python-multipart
pytz
argon2-cffi
authx
regex
# pip install -r requirements.txt